misc/HtmlTestSuite/results/test-1116.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="resources/utils.js"/>

</head><body><iframe sandbox="" srcdoc="&lt;input autofocus&gt;"/>

<script>
'use strict';

promise_test(async () =&gt; {
  await waitForLoad(window);
  await waitUntilStableAutofocusState();
  assert_not_equals(document.activeElement, document.querySelector('iframe'));
}, 'If the sandboxed automatic features browsing context flag is set, ' +
    'autofocus in the browsing context should not be handled.');
</script>
</body></html>
